Add support for multi data-point selection.
I tried to bind ISelectionId using `data(selectionId)`, but then
`ISelectionManager.select()` started failing, so I had to calculate
the `ISelectionId` later on the click handler before calling `select()`.

Another weird thing is the fact that `ids` is of type `extensibility.ISelectionId`
and I had to cast to `visuals.ISelectionId`.